refactor: app and client layers segregration for balance, renames (#167)
