+++ src/app/operations/mod.rs
@@ -1,7 +1,5 @@
 pub mod auth;
-pub mod default_wallet;
 pub mod globals;
-pub mod intraledger_payment;
-pub mod me;
-pub mod set_username;
-pub mod wallet_balance;
+pub mod intraledger;
+pub mod user;
+pub mod wallet;
