--- src/app/operations/wallet_balance.rs
@@ -1,24 +1,0 @@
-use anyhow::{Context, Result};
-
-use crate::app::App;
-
-impl App {
-    pub async fn wallet_balance(
-        &self,
-        btc: bool,
-        usd: bool,
-        wallet_ids: Vec<String>,
-    ) -> Result<()> {
-        let balances = self
-            .client
-            .fetch_balance(btc, usd, wallet_ids)
-            .await
-            .context("Error occurred while fetching wallet balances")?;
-
-        let balances_json =
-            serde_json::to_string_pretty(&balances).context("Can't serialize json")?;
-
-        println!("{}", balances_json);
-        Ok(())
-    }
-}
